scripts/lamps/requirements.txt
phue==1.1
python-dotenv==0.20.0
